flowedit: Move canvas area view rendering to canvas_view::view_canvas_area (#2597)
Extract the canvas widget creation, zoom controls, tooltip overlay,
initializer editor dialog, and context menu overlay from FlowEdit::view()
into a standalone view_canvas_area() function in canvas_view.rs. This
reduces main.rs by ~212 lines with no behavior changes.
